I doodled all over this main class here. Never mind the potentially crass language - the references to 'we' and eldritch horrors are me talking to the code. It's what I do - I put this WIP here in case you come back in the middle of the night to check on the TODO's. Also: There needs to be threading for all the load/store/whatever operations (ExecutorService exists for this) so that the GUI will stop blocking every time something happens.
Please note that everything is probably horrifically broken while awaiting
some of the segregation that I had in mind. It's a pretty classic
system, so I hope it is intuitive from the packages.
The reasons for doing this are maintainability and extensibility. I
realise it's normally kind of retarded to optimise for the future, but
I've found repeatedly that a little segregation early on goes a long way
later.